Create API for netcontrol broker plugin
This allows the use of the netcontrol broker plugin with just a handful
of commands, shielding away the complexity of using broker directly.

After creating a listener by issuing something akin:

ep = netcontrol.Endpoint("bro/event/netcontrol-example", "127.0.0.1", 9977);

a call to ep.getNextCommand() gets the next command with the data
structures formatted in a pyton dict.